Process exit improvements
* Check if process is running before killing it.
* Grab the output from the output variable instead of trying to read the stream again.

+++ lib/PuppeteerSharp/Launcher.cs
@@ -46,7 +46,7 @@ namespace PuppeteerSharp
         private LaunchOptions _options;
         private TaskCompletionSource<bool> _waitForChromeToClose;
         private static int _processCount = 0;
-
+        private bool _processLoaded;
         private const string UserDataDirArgument = "--user-data-dir";
         #endregion
 
@@ -156,6 +156,7 @@ namespace PuppeteerSharp
 
                 if (options.LogProcess)
                 {
+                    _processLoaded = true;
                     Console.WriteLine($"PROCESS COUNT: {Interlocked.Increment(ref _processCount)}");
                 }
 
@@ -164,7 +165,7 @@ namespace PuppeteerSharp
             catch (Exception ex)
             {
                 ForceKillChrome();
-                throw new Exception("Failed to create connection", ex);
+                throw new ChromeProcessException("Failed to create connection", ex);
             }
 
         }
@@ -236,10 +237,14 @@ namespace PuppeteerSharp
 
             EventHandler exitedEvent = (sender, e) =>
             {
+                if (_options.LogProcess && !_processLoaded)
+                {
+                    Console.WriteLine($"PROCESS COUNT: {Interlocked.Increment(ref _processCount)}");
+                }
+
                 CleanUp();
 
-                var error = chromeProcess.StandardError.ReadToEnd();
-                taskWrapper.SetException(new ChromeProcessException($"Failed to launch chrome! {error}"));
+                taskWrapper.SetException(new ChromeProcessException($"Failed to launch chrome! {output}"));
             };
 
             chromeProcess.ErrorDataReceived += (sender, e) =>
@@ -340,7 +345,7 @@ namespace PuppeteerSharp
         {
             try
             {
-                if (_chromeProcess.Id != 0 && Process.GetProcessById(_chromeProcess.Id) != null)
+                if (_chromeProcess.Id != 0 && !_chromeProcess.HasExited && Process.GetProcessById(_chromeProcess.Id) != null)
                 {
                     _chromeProcess.Kill();
                     _chromeProcess.WaitForExit();
